code cleanup :D
